adding some functionalities

+++ app/Http/Controllers/ChirpController.php
@@ -17,14 +17,15 @@ class ChirpController extends Controller
  
  public function store(Request $request){
    $validated=$request ->validate([
-      'message'=>'required|string|max:300'
+      'message'=>'required|string|max:255'
    ]);
 $request->user()->chirps()->create($validated);
    return redirect(route('dashboard'));
  }
 
  public function destroy(Chirp $chirp){
-  if($chirp->user_id!==auth()->id()){
+  Gate::authorize('delete', $chirp);
+  if($chirp->user_id !== auth()->id()){
     abort(403, 'This is not your chirp to delete!');
   }
    $chirp->delete();
@@ -39,9 +40,15 @@ $request->user()->chirps()->create($validated);
  public function update(Request $request, Chirp $chirp){
   Gate::authorize('update', $chirp);
   $validated=$request ->validate([
-    'message'=>'required|string|max:300'
+    'message'=>'required|string|max:255'
   ]);
   $chirp->update($validated);
   return redirect(route('dashboard'));
+ }
+ public function search(Request $request){
+  $chirps=Chirp::with('user')->where('message','like','%'.$request->search.'%')->get();
+  return view('chirps.search',[
+    'chirps'=>$chirps
+  ]);
  }
 }
